move artifact fetching logic to common
